chore: Use Istio slice helpers
